#GREY-329 (hb-develop) Fixed error mesages when deleting (subjects ui)

+++ backend/app/controllers/uoatags.js
@@ -4,6 +4,7 @@ var client = require('app/db_bootstrap'),
     BoLogger = require('app/bologger'),
     bologger = new BoLogger(),
     UnitOfAnalysisTag = require('app/models/uoatags'),
+    UnitOfAnalysisTagLink = require('app/models/uoataglinks'),
     AccessMatrix = require('app/models/access_matrices'),
     Translation = require('app/models/translations'),
     Language = require('app/models/languages'),
@@ -100,6 +101,10 @@ module.exports = {
     deleteOne: function (req, res, next) {
         var thunkQuery = req.thunkQuery;
         co(function* () {
+            var result = yield thunkQuery(UnitOfAnalysisTagLink.select().where(UnitOfAnalysisTagLink.uoaTagId.equals(req.params.id)));
+            if (_.first(result)) {
+                throw new HttpError(403, 'Tag used in Subject to Tag link. Could not delete Tag');
+            }
             return yield thunkQuery(UnitOfAnalysisTag.delete().where(UnitOfAnalysisTag.id.equals(req.params.id)));
         }).then(function () {
             bologger.log({
